services/youtube/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
google-api-python-client==2.154.0
google-auth==2.37.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
yt-dlp
